Problem: warning C4244: 'initializing': conversion from 'double' to 'unsigned int', possible loss of data (#1312)
Solution: add explicit typecasting

+++ src/lib/OpenEXR/ImfInputFile.cpp
@@ -352,12 +352,10 @@ bufferedReadPixels (InputFile::Data* ifd, int scanLine1, int scanLine2)
 
                     switch ( toSlice.type)
                     {
-                        case UINT:
-                        {
-                            unsigned int fill = toSlice.fillValue;
-                            for (int x = xStart;
-                                x <= levelRange.max.x;
-                                x += toSlice.xSampling)
+                    case UINT: {
+                        unsigned int fill = static_cast<unsigned int>(toSlice.fillValue);
+                        for (int x = xStart; x <= levelRange.max.x;
+                             x += toSlice.xSampling)
                             {
                                 * reinterpret_cast<unsigned int*>(toPtr) = fill;
                                 toPtr += toSlice.xStride;
